refactor: move visualization main into own module, write svg in one go
