+++ clusterclue/visualize/molecule.py
@@ -1,17 +1,5 @@
-import csv
-from collections import defaultdict
 from rdkit import Chem
 from rdkit.Chem import Draw
-
-
-def read_compounds(compounds_filepath):
-    compounds = defaultdict(list)
-    with open(compounds_filepath, "r") as f:
-        reader = csv.reader(f, delimiter="\t")
-        for row in reader:
-            bgc_id, name, smiles = row[0], row[1], row[2]
-            compounds[bgc_id].append((name, smiles))
-    return compounds
 
 
 def draw_compounds(compounds):
